+++ packages/xc-cfg/src/builders/extrinsics/xcm/index.ts
@@ -1,6 +1,6 @@
 export { ethereumXcm } from './ethereumXcm';
 export { polkadotXcm } from './polkadotXcm';
-export { xTransfer } from './/xTransfer';
+export { xTokens } from './xTokens';
 
 export * from './builder';
 export * from './types';