+++ packages/xc-cfg/src/chains/polkadot/crust.ts
@@ -25,7 +25,7 @@ export const crust = new Parachain({
   name: 'Crust network',
   parachainId: 2008,
   ss58Format: 88,
-  usesSdkCompat: true,
+  usesLegacyEnhancer: true,
   ws: [
     'wss://crust-parachain.crustnetwork.app',
     'wss://crust-parachain.crustapps.net',