+++ packages/xc-core/src/chain/Parachain.ts
@@ -60,7 +60,7 @@ export interface ParachainParams extends ChainParams<ParachainAssetData> {
   usesDeliveryFee?: boolean;
   usesSignerFee?: boolean;
   usesH160Acc?: boolean;
-  usesSdkCompat?: boolean;
+  usesLegacyEnhancer?: boolean;
   ws: string | string[];
   xcmVersion?: XcmVersion;
 }
@@ -86,7 +86,7 @@ export class Parachain extends Chain<ParachainAssetData> {
 
   readonly usesH160Acc: boolean;
 
-  readonly usesSdkCompat: boolean;
+  readonly usesLegacyEnhancer: boolean;
 
   readonly ws: string | string[];
 
@@ -102,7 +102,7 @@ export class Parachain extends Chain<ParachainAssetData> {
     usesDeliveryFee = false,
     usesSignerFee = false,
     usesH160Acc = false,
-    usesSdkCompat = false,
+    usesLegacyEnhancer = false,
     ws,
     xcmVersion = XcmVersion.v4,
     ...others
@@ -117,14 +117,14 @@ export class Parachain extends Chain<ParachainAssetData> {
     this.usesDeliveryFee = usesDeliveryFee;
     this.usesSignerFee = usesSignerFee;
     this.usesH160Acc = usesH160Acc;
-    this.usesSdkCompat = usesSdkCompat;
+    this.usesLegacyEnhancer = usesLegacyEnhancer;
     this.ws = ws;
     this.xcmVersion = xcmVersion;
   }
 
   get client(): PolkadotClient {
     const pool = SubstrateApis.getInstance();
-    return pool.api(this.ws, undefined, this.usesSdkCompat);
+    return pool.api(this.ws, undefined, this.usesLegacyEnhancer);
   }
 
   getType(): ChainType {